Restructure source structure, place nodes under src/nodes
