+++ src/nodes/shaders/bsdf.c
@@ -9,6 +9,6 @@
 #include "../../includes.h"
 #include "../../datatypes/vector.h"
 #include "../../datatypes/color.h"
-#include "../samplers/sampler.h"
+#include "../../renderer/samplers/sampler.h"
 #include "../../datatypes/material.h"
 #include "bsdf.h"
