+++ src/nodes/shaders/diffuse.h
