+++ src/nodes/shaders/mix.c
@@ -8,7 +8,7 @@
 
 #include "../../includes.h"
 #include "../../datatypes/color.h"
-#include "../samplers/sampler.h"
+#include "../../renderer/samplers/sampler.h"
 #include "../../datatypes/vector.h"
 #include "../../datatypes/material.h"
 #include "bsdf.h"
