+++ src/nodes/textures/texturenode.h
