+++ src/renderer/pathtrace.h
@@ -12,7 +12,7 @@
 #include "../datatypes/poly.h"
 #include "../datatypes/lightRay.h"
 #include "../datatypes/material.h"
-#include "bsdf/bsdf.h"
+#include "../nodes/shaders/bsdf.h"
 
 struct world;
